requirements.txt
alembic==1.13.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
coverage==7.4.4
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
py==1.11.0
pytest==6.2.5
pytest-cov==5.0.0
pytest-pythonpath==0.7.4
python-dateutil==2.9.0.post0
six==1.16.0
SQLAlchemy==2.0.29
toml==0.10.2
typing_extensions==4.11.0
Werkzeug==3.0.2
